back/requirements.txt
markdown
django-filter
Django==4.1.2
channels==4.0.0
channels-redis==4.0.0
celery==5.2.7
django-celery-results
django-celery-beat
djangorestframework==3.14.0
djangorestframework-dataclasses
drf-spectacular[sidecar]
django_nextjs
openai
ipython
django-jazzmin
psycopg2-binary
uvicorn[standard]